requirements.txt
graphiti-core>=0.18.0
kafka-python
neo4j
openai
agno>=1.1.6
fastapi
uvicorn[standard]
networkx
torch
torch-geometric
pydantic
aiokafka
